+++ Sources/UILabelExtensions.swift
@@ -56,16 +56,11 @@ extension UILabel {
         self.fitHeight()
         sizeToFit()
     }
-
-    /// EZSwiftExtensions
-    public func setText(_ text: String?, animated: Bool, duration: TimeInterval?) {
-        if animated {
-            UIView.transition(with: self, duration: duration ?? 0.3, options: .transitionCrossDissolve, animations: { () -> Void in
-                self.text = text
-                }, completion: nil)
-        } else {
-            self.text = text
-        }
-
+    
+    /// EZSwiftExtensions (if duration set to 0 animate wont be)
+    public func set(text _text: String?, duration: TimeInterval) {
+        UIView.transition(with: self, duration: duration, options: .transitionCrossDissolve, animations: { () -> Void in
+            self.text = _text
+        }, completion: nil)
     }
 }